moved data text input to it's own file
