+++ frontend-builder/src/components/Shared/TabPanel/index.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import Tabs from '@mui/material/Tabs'
-import Tab from '@mui/material/Tab'
 import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
 interface TabPanelProps {
